Correct Spelling mistake
Line 4 from "This is the Quoter BerrerDiscord plugin help page. Most likely you clicked on the quote of another user who used this plugin in the Discord messenger. You do not have this plugin installed, so you are here. If the plug-in was installed, you would jump to the quoted message." to "This is the Quoter BetterDiscord plugin help page. Most likely you clicked on the quote of another user who used this plugin in the Discord messenger. You do not have this plugin installed, so you are here. If the plug-in was installed, you would jump to the quoted message."